Organized files. Now on to write tests and publish on rubygems.org
