+++ .rspec
@@ -1,0 +1,2 @@
+--format documatation
+--color
