+++ Gemfile
@@ -1,0 +1,4 @@
+source 'https://rubygems.org'
+
+# Specify dependencies in wiki_on_this_day.gemspec
+gemspec
