+++ lib/wiki_on_this_day/version.rb
@@ -1,0 +1,3 @@
+module WikiOnThisDay
+  VERSION = "0.1.0"
+end
